pyproject.toml
[tool.poetry]
name = "whatalesyou"
version = "0.1.0"
description = ""
authors = ["Bill Willits <willitsw@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.5"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
gunicorn = "^20.1.0"
whitenoise = "^6.3.0"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^3.13.0"
black = "^23.1.0"
isort = "^5.12.0"
pre-commit = "^3.0.4"
drf-writable-nested = "^0.7.0"
debugpy = "^1.8.0"
django-webpack-loader = "^2.0.1"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.4"
pep8 = "^1.7.1"
pylint = "^3.0.2"
pylint-django = "^2.5.5"

[build-system]
requires = ["poetry-core", "poetry>=1.3.2"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
